requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
qdrant-client==1.12.0
openai==1.55.0
cohere==5.11.0
psycopg2-binary==2.9.10
httpx==0.27.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
PyJWT==2.8.0
email-validator==2.1.0